% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{ggimg_util_load}
\alias{ggimg_util_load}
\title{Utility Function to Load Image Arrays}
\usage{
ggimg_util_load(path)
}
\arguments{
\item{path}{character vector giving the path to the images}
}
\description{
Can be used within a pipeline of dplyr verbs to load images as a list
column. Images that cannot be loaded will be returned as \code{NULL}
rather than producing an error.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
