% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{gimg_load_img}
\alias{gimg_load_img}
\title{Load an image}
\usage{
gimg_load_img(input)
}
\arguments{
\item{input}{the input as either a path, url, or the image itself}
}
\value{
An array with the image.
}
\description{
Given an input, either a URL, local file path, or a RasterImage itself,
return an array of the image file.
}
