% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{posters}
\alias{posters}
\title{Movie Posters from Animated Films}
\description{
Data frame containing information about the 50 top grossing animated U.S.
films from 2000-2019. The column points to a thumbnail of the movie's
promotional poster, installed along with the package. See \code{geom_img}
for an example of using it with the package.
}
